% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_exercise.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on a state map for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer giving the state number (matching the `STATE`
column in the FARS dataset).}

\item{year}{A numeric year indicating which year's data to use.}
}
\value{
A map is drawn as a side effect. If no accidents are present for
the specified state,the function returns `NULL` invisibly.
}
\description{
This function reads FARS data for a specified year and plots accidents for a
given state on a U.S. map. If no accidents are found for the state, a
message is shown.
}
\details{
Longitude values greater than 900 and latitude values greater than 90 are
set to `NA`prior to plotting, since they are invalid.
}
\examples{
fars_map_state(1, 2013)

}
