% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_exercise.R
\name{fars_read}
\alias{fars_read}
\title{Read in the fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename of the fars data file.}
}
\value{
A table-like dataframe containing the contents of the FARS data file
}
\description{
This function checks wether the provided filename exists, if it does not it
prints that the file does not exist and stops the function. If the file does
exist, the data is read without returning messages or showing the progress and
the data is then made into a table-dataframe using dplyr.
}
\examples{
fars_read("fars_data.csv")

}
