% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_exercise.R
\name{make_filename}
\alias{make_filename}
\title{Create a FARS filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or numeric value specifying the year.}
}
\value{
A character string with the FARS filename in the format
"accident_<year>.csv.bz2".
}
\description{
This function generates a standard FARS filename string for a given year.
}
\examples{
make_filename(2015)

}
